backend/app/services/calendar_service.py
"""
Calendar Service for scheduling handover meetings
"""
import uuid
from datetime import datetime
from flask import current_app


class CalendarService:
    """
    Service for managing calendar events.
    This is a mock implementation - replace with actual Google Calendar API integration.
    """
    
    @staticmethod
    def create_event(schedule):
        """
        Create a calendar event for a handover schedule.
        
        For production, integrate with Google Calendar API:
        1. Enable Google Calendar API in Google Cloud Console
        2. Configure OAuth credentials
        3. Use google-api-python-client library
        
        Returns the calendar event ID.
        """
        try:
            # Mock implementation - generates a unique event ID
            event_id = f"evt_{uuid.uuid4().hex[:12]}"
            
            # In production, use Google Calendar API:
            # from googleapiclient.discovery import build
            # service = build('calendar', 'v3', credentials=credentials)
            # 
            # event = {
            #     'summary': schedule.title,
            #     'description': schedule.description,
            #     'start': {
            #         'dateTime': f"{schedule.scheduled_date}T{schedule.start_time}",
            #         'timeZone': 'UTC',
            #     },
            #     'end': {
            #         'dateTime': f"{schedule.scheduled_date}T{schedule.end_time}",
            #         'timeZone': 'UTC',
            #     },
            #     'attendees': [{'email': get_user_email(uid)} for uid in schedule.attendees],
            #     'location': schedule.location,
            #     'conferenceData': {...} if schedule.meeting_link else None,
            # }
            # 
            # result = service.events().insert(calendarId='primary', body=event).execute()
            # return result.get('id')
            
            current_app.logger.info(f"Calendar event created: {event_id} for schedule {schedule.id}")
            return event_id
            
        except Exception as e:
            current_app.logger.error(f"Failed to create calendar event: {str(e)}")
            return None
    
    @staticmethod
    def update_event(schedule):
        """
        Update an existing calendar event.
        
        Returns True if successful, False otherwise.
        """
        try:
            if not schedule.calendar_event_id:
                return False
            
            # Mock implementation
            # In production, use:
            # service.events().update(
            #     calendarId='primary',
            #     eventId=schedule.calendar_event_id,
            #     body=updated_event
            # ).execute()
            
            current_app.logger.info(f"Calendar event updated: {schedule.calendar_event_id}")
            return True
            
        except Exception as e:
            current_app.logger.error(f"Failed to update calendar event: {str(e)}")
            return False
    
    @staticmethod
    def delete_event(schedule):
        """
        Delete a calendar event.
        
        Returns True if successful, False otherwise.
        """
        try:
            if not schedule.calendar_event_id:
                return False
            
            # Mock implementation
            # In production, use:
            # service.events().delete(
            #     calendarId='primary',
            #     eventId=schedule.calendar_event_id
            # ).execute()
            
            current_app.logger.info(f"Calendar event deleted: {schedule.calendar_event_id}")
            return True
            
        except Exception as e:
            current_app.logger.error(f"Failed to delete calendar event: {str(e)}")
            return False
    
    @staticmethod
    def get_event(event_id):
        """
        Get calendar event details.
        
        Returns event data or None if not found.
        """
        try:
            if not event_id:
                return None
            
            # Mock implementation
            # In production, use:
            # event = service.events().get(
            #     calendarId='primary',
            #     eventId=event_id
            # ).execute()
            # return event
            
            return {
                'id': event_id,
                'status': 'confirmed'
            }
            
        except Exception as e:
            current_app.logger.error(f"Failed to get calendar event: {str(e)}")
            return None
    
    @staticmethod
    def list_events(start_date, end_date, user_id=None):
        """
        List calendar events within a date range.
        
        Returns list of events.
        """
        try:
            # Mock implementation
            # In production, use:
            # events_result = service.events().list(
            #     calendarId='primary',
            #     timeMin=start_date.isoformat() + 'Z',
            #     timeMax=end_date.isoformat() + 'Z',
            #     singleEvents=True,
            #     orderBy='startTime'
            # ).execute()
            # return events_result.get('items', [])
            
            return []
            
        except Exception as e:
            current_app.logger.error(f"Failed to list calendar events: {str(e)}")
            return []
    
    @staticmethod
    def send_invite(schedule, attendee_emails):
        """
        Send calendar invites to attendees.
        
        Returns True if successful, False otherwise.
        """
        try:
            # Mock implementation
            # In production, calendar invites are sent automatically
            # when attendees are added to an event
            
            current_app.logger.info(f"Calendar invites sent to: {attendee_emails}")
            return True
            
        except Exception as e:
            current_app.logger.error(f"Failed to send calendar invites: {str(e)}")
            return False


def get_google_calendar_service():
    """
    Get authenticated Google Calendar service.
    
    For production implementation:
    
    from google.oauth2.credentials import Credentials
    from google_auth_oauthlib.flow import Flow
    from googleapiclient.discovery import build
    
    # Load credentials from storage
    creds = Credentials.from_authorized_user_file('token.json', SCOPES)
    
    # Refresh if expired
    if creds and creds.expired and creds.refresh_token:
        creds.refresh(Request())
    
    # Build service
    service = build('calendar', 'v3', credentials=creds)
    return service
    """
    pass
