backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.36
Werkzeug==3.0.1
python-dotenv==1.0.0
PyJWT==2.8.0
gunicorn==21.2.0

# PostgreSQL driver (psycopg3 for Python 3.13+)
psycopg[binary]

# Google OAuth (optional)
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
